backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
boto3==1.35.0
python-dotenv==1.0.1
mangum==0.19.0
opensearch-py==2.4.2
